requirements.txt
# Core
discord.py~=2.7
asyncpg~=0.31
python-dotenv~=1.2

# AI
google-genai~=1.68

# Web/HTTP
aiohttp~=3.13
requests~=2.32

# Parsing & Data
beautifulsoup4~=4.14
lxml~=6.0
langdetect~=1.0

# Image
Pillow~=12.1
imgkit~=1.2

# Utilities
demoji~=1.1
yarl~=1.23
Reverso-API
